GRAF-91 Upload photo auth guard & other minor changes
